Rename Woodcutting folder to correct woodcutting (#5829)
