+++ src/lib/skilling/skills/woodcutting/woodcutting.ts
