+++ src/tasks/minions/woodcuttingActivity.ts
@@ -6,8 +6,8 @@ import { MediumSeedPackTable } from '../../lib/data/seedPackTables';
 import addSkillingClueToLoot from '../../lib/minions/functions/addSkillingClueToLoot';
 import { eggNest } from '../../lib/simulation/birdsNest';
 import { soteSkillRequirements } from '../../lib/skilling/functions/questRequirements';
-import { ForestryEvents, LeafTable } from '../../lib/skilling/skills/Woodcutting/forestry';
-import Woodcutting from '../../lib/skilling/skills/Woodcutting/woodcutting';
+import { ForestryEvents, LeafTable } from '../../lib/skilling/skills/woodcutting/forestry';
+import Woodcutting from '../../lib/skilling/skills/woodcutting/woodcutting';
 import { SkillsEnum } from '../../lib/skilling/types';
 import { WoodcuttingActivityTaskOptions } from '../../lib/types/minions';
 import { perTimeUnitChance, roll, skillingPetDropRate } from '../../lib/util';
